Install packages on context load
This might not always be desirable, so might get tuned in a future
version to allow a command line option to disable (and a function
argument).  However, it's going to be essential for the cluster use.

This required a bit of replumbing how the local libraries are set up,
and assorted workarounds for R CMD check and package installation.