Rework the bootstrap again
Now install context from a local drat that is saved when the context
is created.  That's obviously a little heavy, so I'll need to limit
how often it is downloaded.  Once it switches over to being versioned
this will be easy as now we'll refuse to redownload a package that
points at a versioned reference.

The local drat is therefore _always_ used, which removes a few
conditionals.

I'm caching downloads, though not sure how useful that will be in
practice.  It should make tests a little faster though.