Record failure on context startup
This captures all failures that happen from the point where the
context database has failed (which is going to be the bulk of
failures, especially when it's due to things like missing source
files, unstated dependencies, etc).

What this can't catch is failures due to inability to connect to
the storage, etc; there's really no nice way of dealing with that.